+++ test/multiconductor.jl
@@ -41,6 +41,42 @@ end
 
 @testset "test multi-conductor" begin
 
+    @testset "json parser" begin
+        mc_data = build_mc_data("../test/data/pti/parser_test_defaults.raw")
+        mc_json_string = PowerModels.parse_json(JSON.json(mc_data))
+        if VERSION > v"0.7.0-"
+            @test mc_data == mc_json_string
+        end
+
+        io = PipeBuffer()
+        JSON.print(io, mc_data)
+        mc_json_file = PowerModels.parse_file(io)
+        if VERSION > v"0.7.0-"
+            @test mc_data == mc_json_file
+        else
+            @test InfrastructureModels.compare_dict(mc_data, mc_json_file)
+        end
+
+        mc_strg_data = build_mc_data("../test/data/matpower/case5_strg.m")
+        mc_strg_json_string = PowerModels.parse_json(JSON.json(mc_strg_data))
+        if VERSION > v"0.7.0-"
+            @test mc_strg_data == mc_strg_json_string
+        else
+            @test InfrastructureModels.compare_dict(mc_strg_data, mc_strg_json_string)
+        end
+
+        # test that non-multiconductor json still parses, pti_json_file will result in error if fails
+        pti_data = PowerModels.parse_file("../test/data/pti/parser_test_defaults.raw")
+        io = PipeBuffer()
+        JSON.print(io, pti_data)
+        pti_json_file = PowerModels.parse_file(io)
+        if VERSION > v"0.7.0-"
+            @test pti_data == pti_json_file
+        else
+            @test InfrastructureModels.compare_dict(pti_data, pti_json_file)
+        end
+    end
+
     @testset "idempotent unit transformation" begin
         @testset "5-bus replicate case" begin
             mp_data = build_mc_data("../test/data/matpower/case5_dc.m")
@@ -471,8 +507,8 @@ end
         @test PowerModels.conductors(c) == 3
         @test PowerModels.conductors(a) == 3
         @test all(size(a) == (3,3))
-        @test isa(JSON.lower(a), Array)
-        @test all(JSON.lower(a) == a.values)
+        @test isa(JSON.lower(a), Dict)
+        @test all(JSON.lower(a)["values"] == a.values)
         @test !isapprox(d, e)
         @test PowerModels.getmcv(a, 1, 1) == a[1,1]